--- src/components/formik/FormikField/SingleSelect/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./SingleSelect";
-export * from "./SingleSelect";
