--- src/components/formik/FormikField/StatefulToggleField/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./StatefulToggleField";
-export * from "./StatefulToggleField";
