--- src/components/formik/FormikField/TextArea/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./TextArea";
-export * from "./TextArea";
