--- src/components/formik/FormikField/TextField/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./TextField";
-export * from "./TextField";
