--- src/components/formik/FormikField/ToggleField/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ToggleField";
-export * from "./ToggleField";
