--- src/components/formik/FormikField/UploadFileField/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./UploadFileField";
-export * from "./UploadFileField";
