+++ src/components/formik/FormikField/index.ts
@@ -1,29 +1,12 @@
-import SingleSelect from "./SingleSelect";
-import type { SingleSelectProps } from "./SingleSelect";
-import TextField from "./TextField";
-import type { TextFieldProps } from "./TextField";
-import UploadFileField from "./UploadFileField";
-import type { UploadFileFieldProps } from "./UploadFileField";
-import TextArea, { TextAreaProps } from "./TextArea";
-import ToggleField from "./ToggleField";
-import type { ToggleFieldProps } from "./ToggleField";
-import StatefulToggleField from "./StatefulToggleField";
-import type { StatefulToggleFieldProps } from "@instill-ai/design-system";
-
-export {
-  SingleSelect,
-  TextField,
-  UploadFileField,
-  TextArea,
-  ToggleField,
-  StatefulToggleField,
-};
-
-export type {
-  SingleSelectProps,
-  TextFieldProps,
-  UploadFileFieldProps,
-  TextAreaProps,
-  ToggleFieldProps,
-  StatefulToggleFieldProps,
-};
+export { SingleSelect } from "./SingleSelect";
+export type { SingleSelectProps } from "./SingleSelect";
+export { TextField } from "./TextField";
+export type { TextFieldProps } from "./TextField";
+export { UploadFileField } from "./UploadFileField";
+export type { UploadFileFieldProps } from "./UploadFileField";
+export { TextArea } from "./TextArea";
+export type { TextAreaProps } from "./TextArea";
+export { ToggleField } from "./ToggleField";
+export type { ToggleFieldProps } from "./ToggleField";
+export { StatefulToggleField } from "./StatefulToggleField";
+export type { StatefulToggleFieldProps } from "./StatefulToggleField";
