--- src/components/formik/FormikFormBase/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./FormikFormBase";
-export * from "./FormikFormBase";
