+++ src/components/formik/index.ts
@@ -1,40 +1,5 @@
-import {
-  SingleSelect,
-  TextArea,
-  TextField,
-  ToggleField,
-  UploadFileField,
-} from "./FormikField";
+export { FormikFormBase } from "./FormikFormBase";
+export type { FormikFormBaseProps } from "./FormikFormBase";
 
-import type {
-  SingleSelectProps,
-  TextAreaProps,
-  TextFieldProps,
-  ToggleFieldProps,
-  UploadFileFieldProps,
-} from "./FormikField";
-
-import { FormikStep, FormikMultiStep } from "./FormikMultiStep";
-import type { FormikMultiStepProps } from "./FormikMultiStep";
-import FormikFormBase from "./FormikFormBase";
-import type { FormikFormBaseProps } from "./FormikFormBase";
-
-export {
-  SingleSelect,
-  TextArea,
-  TextField,
-  ToggleField,
-  UploadFileField,
-  FormikStep,
-  FormikMultiStep,
-  FormikFormBase,
-};
-export type {
-  SingleSelectProps,
-  TextAreaProps,
-  TextFieldProps,
-  ToggleFieldProps,
-  UploadFileFieldProps,
-  FormikMultiStepProps,
-  FormikFormBaseProps,
-};
+export * from "./FormikField";
+export * from "./FormikMultiStep";
